refactor: centralize config paths and health checks
