+++ backend/main.py
@@ -8,18 +8,25 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
 import uvicorn
 import os
-import logging
 import secrets
 import json
 from datetime import datetime
 from typing import Dict, Any
+import logging
+
+from utils.logging_config import setup_logging, get_logger
+from utils.responses import create_standardized_response, create_error_response
+
+# configure logging before other services use it
+setup_logging()
+logger = get_logger(__name__)
 
 # Import database models
 from models import db_manager, Session, Submission, Evaluation
 
 # Import services
 from services import (
-    config_service, 
+    config_service,
     get_auth_service,
     authenticate,
     validate_session,
@@ -38,9 +45,8 @@ from services.llm_service import EnhancedLLMService
 # Import authentication decorators
 from decorators import require_auth
 
-# Configure logging (will be updated after config loading)
-logging.basicConfig(level=logging.DEBUG, format='%(levelname)s:%(name)s:%(message)s')
-logger = logging.getLogger(__name__)
+# Import routers
+from routers.health import router as health_router
 
 # Function to update logging level based on config
 def update_logging_level():
@@ -62,33 +68,6 @@ def update_logging_level():
     except Exception as e:
         logger.error(f"Failed to update logging level: {e}")
 
-# Helper functions for standardized responses
-def create_standardized_response(data: Any, status_code: int = 200) -> Dict[str, Any]:
-    """Create a standardized API response with {data, meta, errors} format"""
-    return {
-        "data": data,
-        "meta": {
-            "timestamp": datetime.utcnow().isoformat(),
-            "request_id": "placeholder"
-        },
-        "errors": []
-    }
-
-def create_error_response(error_code: str, message: str, field: str = None, details: str = None, status_code: int = 400) -> Dict[str, Any]:
-    """Create a standardized error response"""
-    return {
-        "data": None,
-        "meta": {
-            "timestamp": datetime.utcnow().isoformat(),
-            "request_id": "placeholder"
-        },
-        "errors": [{
-            "code": error_code,
-            "message": message,
-            "field": field,
-            "details": details
-        }]
-    }
 
 # Create FastAPI app
 app = FastAPI(
@@ -122,6 +101,9 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
+# Include routers
+app.include_router(health_router)
+
 @app.get("/")
 async def root():
     """Root endpoint"""
@@ -131,329 +113,6 @@ async def root():
         "status": "running"
     }
 
-@app.get("/health")
-async def health_check():
-    """Health check endpoint"""
-    try:
-        # Check database health
-        db_health = db_manager.health_check()
-        
-        # Check configuration health
-        config_health = config_service.health_check()
-        
-        # Check LLM health
-        try:
-            llm_service = EnhancedLLMService()
-            llm_health = llm_service.validate_configuration()
-            llm_status = "healthy" if llm_health["valid"] else "unhealthy"
-        except Exception as e:
-            logger.error(f"LLM health check failed: {e}")
-            llm_status = "unhealthy"
-        
-        # Check authentication health
-        try:
-            auth_service = get_auth_service()
-            auth_health = auth_service.health_check()
-            auth_status = auth_health["status"]
-        except Exception as e:
-            logger.error(f"Auth health check failed: {e}")
-            auth_status = "unhealthy"
-        
-        # Basic health check
-        health_status = {
-            "status": "healthy",
-            "timestamp": datetime.utcnow().isoformat(),
-            "version": "1.0.0",
-            "services": {
-                "api": "healthy",
-                "database": db_health["status"],
-                "configuration": config_health["status"],
-                "llm": llm_status,
-                "auth": auth_status
-            }
-        }
-        
-        # Note: Detailed health information moved to authenticated endpoints
-        # This public endpoint provides only basic service status
-        
-        # Check if any service is unhealthy
-        if any(status != "healthy" for status in health_status["services"].values()):
-            health_status["status"] = "unhealthy"
-            return JSONResponse(
-                status_code=503, 
-                content=create_standardized_response(health_status)
-            )
-        
-        return create_standardized_response(health_status)
-        
-    except Exception as e:
-        logger.error(f"Health check failed: {e}")
-        return JSONResponse(
-            status_code=503,
-            content=create_error_response(
-                "HEALTH_CHECK_FAILED",
-                "Health check failed",
-                details=str(e)
-            )
-        )
-
-@app.get("/health/detailed")
-@require_auth(admin_only=True)
-async def detailed_health_check(request: Request):
-    """Detailed health check endpoint - admin only with full system information"""
-    try:
-        # Check database health
-        db_health = db_manager.health_check()
-        
-        # Check configuration health
-        config_health = config_service.health_check()
-        
-        # Check LLM health
-        try:
-            llm_service = EnhancedLLMService()
-            llm_health = llm_service.validate_configuration()
-            llm_status = "healthy" if llm_health["valid"] else "unhealthy"
-        except Exception as e:
-            logger.error(f"LLM health check failed: {e}")
-            llm_status = "unhealthy"
-        
-        # Check authentication health
-        try:
-            auth_service = get_auth_service()
-            auth_health = auth_service.health_check()
-            auth_status = auth_health["status"]
-        except Exception as e:
-            logger.error(f"Auth health check failed: {e}")
-            auth_status = "unhealthy"
-        
-        # Detailed health check with full information
-        health_status = {
-            "status": "healthy",
-            "timestamp": datetime.utcnow().isoformat(),
-            "version": "1.0.0",
-            "services": {
-                "api": "healthy",
-                "database": db_health["status"],
-                "configuration": config_health["status"],
-                "llm": llm_status,
-                "auth": auth_status
-            }
-        }
-        
-        # Add database details if available
-        if db_health["status"] == "healthy":
-            health_status["database_details"] = {
-                "tables": db_health.get("tables", []),
-                "journal_mode": db_health.get("journal_mode", ""),
-                "user_count": db_health.get("user_count", 0)
-            }
-        else:
-            health_status["database_error"] = db_health.get("error", "Unknown error")
-        
-        # Add configuration details if available
-        if config_health["status"] == "healthy":
-            health_status["config_details"] = {
-                "configs_loaded": config_health.get("configs_loaded", []),
-                "last_loaded": config_health.get("last_loaded", ""),
-                "config_dir": config_health.get("config_dir", "")
-            }
-        else:
-            health_status["config_error"] = config_health.get("error", "Unknown error")
-        
-        # Add LLM details if available
-        if llm_status == "healthy":
-            health_status["llm_details"] = {
-                "provider": "anthropic",
-                "model": llm_health.get("model", "claude-3-haiku-20240307"),
-                "api_accessible": llm_health.get("components", {}).get("claude_client", False),
-                "config_loaded": llm_health.get("components", {}).get("prompt_config", False),
-                "supported_languages": llm_health.get("supported_languages", []),
-                "default_language": llm_health.get("default_language", "en")
-            }
-        else:
-            health_status["llm_error"] = llm_health.get("error", "Unknown error")
-        
-        # Add authentication details if available
-        if auth_status == "healthy":
-            health_status["auth_details"] = {
-                "config_loaded": auth_health.get("config_loaded", False),
-                "active_sessions": auth_health.get("active_sessions", 0),
-                "brute_force_protection": auth_health.get("brute_force_protection", False)
-            }
-        else:
-            health_status["auth_error"] = auth_health.get("error", "Unknown error")
-        
-        # Add environment and debug mode information
-        app_env = os.environ.get('APP_ENV', 'production')
-        
-        # Get debug mode from configuration service (YAML settings)
-        auth_config = config_service.get_auth_config()
-        debug_mode = False
-        if auth_config and 'security_settings' in auth_config:
-            debug_mode = auth_config['security_settings'].get('debug_mode', False)
-        
-        health_status["environment"] = {
-            "app_env": app_env,
-            "debug_mode": debug_mode,
-            "mode": "development" if app_env == "development" or debug_mode else "production"
-        }
-        
-        # Check if any service is unhealthy
-        if any(status != "healthy" for status in health_status["services"].values()):
-            health_status["status"] = "unhealthy"
-            return JSONResponse(
-                status_code=503, 
-                content=create_standardized_response(health_status)
-            )
-        
-        return create_standardized_response(health_status)
-        
-    except Exception as e:
-        logger.error(f"Detailed health check failed: {e}")
-        return JSONResponse(
-            status_code=503,
-            content=create_error_response(
-                "HEALTH_CHECK_FAILED",
-                "Detailed health check failed",
-                details=str(e)
-            )
-        )
-
-@app.get("/health/database")
-@require_auth(admin_only=True)
-async def database_health_check(request: Request):
-    """Database-specific health check endpoint - admin only"""
-    try:
-        db_health = db_manager.health_check()
-        return create_standardized_response({
-            "status": db_health["status"],
-            "timestamp": datetime.utcnow().isoformat(),
-            "database": db_health
-        })
-    except Exception as e:
-        logger.error(f"Database health check failed: {e}")
-        return JSONResponse(
-            status_code=503,
-            content=create_error_response(
-                "DATABASE_HEALTH_FAILED",
-                "Database health check failed",
-                details=str(e)
-            )
-        )
-
-@app.get("/health/config")
-@require_auth(admin_only=True)
-async def config_health_check(request: Request):
-    """Configuration-specific health check endpoint - admin only"""
-    try:
-        config_health = config_service.health_check()
-        return create_standardized_response(
-            {
-                "status": config_health["status"],
-                "timestamp": datetime.utcnow().isoformat(),
-                "configuration": {
-                    "status": config_health["status"],
-                    "configs_loaded": config_health.get("configs_loaded", []),
-                    "last_loaded": config_health.get("last_loaded", ""),
-                    # Remove sensitive config_dir path
-                    "config_count": len(config_health.get("configs_loaded", []))
-                }
-            }
-        )
-    except Exception as e:
-        logger.error(f"Configuration health check failed: {e}")
-        return JSONResponse(
-            status_code=503,
-            content=create_error_response(
-                "CONFIG_HEALTH_FAILED",
-                "Configuration health check failed",
-                details=str(e)
-            )
-        )
-
-@app.get("/health/llm")
-@require_auth(admin_only=True)
-async def llm_health_check(request: Request):
-    """LLM service health check endpoint - admin only"""
-    try:
-        llm_service = EnhancedLLMService()
-        llm_health = llm_service.validate_configuration()
-        
-        # Sanitize sensitive information
-        sanitized_llm_health = {
-            "valid": llm_health["valid"],
-            "components": llm_health.get("components", {}),
-            "supported_languages": llm_health.get("supported_languages", []),
-            "default_language": llm_health.get("default_language", "en"),
-            "model": llm_health.get("model", "unknown")
-        }
-        
-        response_data = {
-            "status": "healthy" if llm_health["valid"] else "unhealthy",
-            "timestamp": datetime.utcnow().isoformat(),
-            "llm": sanitized_llm_health
-        }
-        
-        if llm_health["valid"]:
-            return create_standardized_response(response_data)
-        else:
-            return JSONResponse(
-                status_code=503,
-                content=create_standardized_response(response_data)
-            )
-    except Exception as e:
-        logger.error(f"LLM health check failed: {e}")
-        return JSONResponse(
-            status_code=503,
-            content=create_error_response(
-                "LLM_HEALTH_FAILED",
-                "LLM health check failed",
-                details=str(e)
-            )
-        )
-
-@app.get("/health/auth")
-@require_auth(admin_only=True)
-async def auth_health_check(request: Request):
-    """Authentication service health check endpoint - admin only"""
-    try:
-        auth_service = get_auth_service()
-        auth_health = auth_service.health_check()
-        
-        # Sanitize sensitive authentication information
-        sanitized_auth_health = {
-            "status": auth_health["status"],
-            "service": auth_health.get("service", "authentication"),
-            "config_loaded": auth_health.get("config_loaded", False),
-            "active_sessions": auth_health.get("active_sessions", 0),
-            "brute_force_protection": auth_health.get("brute_force_protection", False),
-            "session_expiry_hours": auth_health.get("session_expiry_hours", 24),
-            "session_token_length": auth_health.get("session_token_length", 32)
-        }
-        
-        response_data = {
-            "status": auth_health["status"],
-            "timestamp": datetime.utcnow().isoformat(),
-            "auth": sanitized_auth_health
-        }
-        
-        if auth_health["status"] == "healthy":
-            return create_standardized_response(response_data)
-        else:
-            return JSONResponse(
-                status_code=503,
-                content=create_standardized_response(response_data)
-            )
-    except Exception as e:
-        logger.error(f"Auth health check failed: {e}")
-        return JSONResponse(
-            status_code=503,
-            content=create_error_response(
-                "AUTH_HEALTH_FAILED",
-                "Authentication health check failed",
-                details=str(e)
-            )
-        )
 
 
 @app.post("/api/v1/auth/logout")
