+++ backend/services/auth_service.py
@@ -6,15 +6,15 @@ Handles admin authentication and session management
 import os
 import bcrypt
 import secrets
-import logging
 from typing import Dict, Any, Optional, Tuple, List
 from datetime import datetime, timedelta
 import yaml
 import hashlib
 
-# Configure logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+from utils.config import get_config_path
+from utils.logging_config import get_logger
+
+logger = get_logger(__name__)
 
 class AuthService:
     """Service for unified user and admin authentication"""
@@ -22,13 +22,8 @@ class AuthService:
     def __init__(self, config_path: str = None):
         """Initialize auth service with automatic path detection"""
         if config_path is None:
-            # In container, config is mounted at /app/config
-            # For development, fallback to ../config
-            if os.path.exists('/app/config'):
-                config_path = '/app/config'
-            else:
-                config_path = '../config'
-        
+            config_path = str(get_config_path())
+
         self.config_path = config_path
         self.auth_config = None
         self.login_attempts = {}  # Track login attempts for brute force protection
